requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.10.0
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.6
WTForms==2.0.1
Werkzeug==0.9.6
argparse==1.2.1
factory-boy==2.4.1
itsdangerous==0.24
py==1.4.22
pytest==2.6.0
wsgiref==0.1.2
